+++ src/__test__/components/NodesSelection.test.js
@@ -1,30 +1,29 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 import toJson from 'enzyme-to-json'
+import { DataSet } from 'vis-data'
 import NodesSelection from '../../components/NodesSelection'
 
 const setup = () => {
   const props = {
     resetSelectedNodes: jest.fn(),
     removeFromArray: jest.fn(),
-    availableNodesNormalised: {
-      abc: {
-        id: 'abc',
-        label: 'label',
-        rdfAbout: 'rdfAbout',
-        skosComment: 'skosComment',
-        skosDefinition: 'skosDefinition',
-        skosExample: 'skosExample'
-      },
-      cde: {
-        id: 'cde',
-        label: 'label2',
-        rdfAbout: 'rdfAbout',
-        skosComment: 'skosComment',
-        skosDefinition: 'skosDefinition',
-        skosExample: 'skosExample'
-      }
+    availableNodes: new DataSet([{
+      id: 'abc',
+      label: 'label',
+      rdfAbout: 'rdfAbout',
+      skosComment: 'skosComment',
+      skosDefinition: 'skosDefinition',
+      skosExample: 'skosExample'
     },
+    {
+      id: 'cde',
+      label: 'label2',
+      rdfAbout: 'rdfAbout',
+      skosComment: 'skosComment',
+      skosDefinition: 'skosDefinition',
+      skosExample: 'skosExample'
+    }]),
     selectedNodes: ['abc']
   }