+++ src/__test__/components/NodesSelectionDetails.test.js
@@ -1,29 +1,29 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 import toJson from 'enzyme-to-json'
+import { DataSet } from 'vis-data'
 import NodesSelectionDetails from '../../components/NodesSelectionDetails'
 
 const setup = () => {
   const props = {
     nodeId: 'abc',
-    availableNodesNormalised: {
-      abc: {
-        id: 'id',
-        label: 'label',
-        rdfAbout: 'rdfAbout',
-        skosComment: 'skosComment',
-        skosDefinition: 'skosDefinition',
-        skosExample: 'skosExample'
-      },
-      cde: {
-        id: 'cde',
-        label: 'label2',
-        rdfAbout: 'rdfAbout',
-        skosComment: 'skosComment',
-        skosDefinition: 'skosDefinition',
-        skosExample: 'skosExample'
-      }
+    availableNodes: new DataSet([{
+      id: 'id',
+      label: 'label',
+      rdfAbout: 'rdfAbout',
+      skosComment: 'skosComment',
+      skosDefinition: 'skosDefinition',
+      skosExample: 'skosExample'
     },
+    {
+      id: 'cde',
+      label: 'label2',
+      rdfAbout: 'rdfAbout',
+      skosComment: 'skosComment',
+      skosDefinition: 'skosDefinition',
+      skosExample: 'skosExample'
+    }
+    ]),
     nodesConnections: [{
       from: 'abc',
       fromLabel: 'label',