+++ src/components/NodesSelectionDetails.js
@@ -5,12 +5,12 @@ import actions from '../store/actions'
 
 const NodesSelectionDetails = ({
   nodeId,
-  availableNodesNormalised,
+  availableNodes,
   nodesConnections,
 }) => {
   const { t } = useTranslation()
 
-  const selectedNode = availableNodesNormalised[nodeId]
+  const selectedNode = availableNodes.get(nodeId)
 
   const tableRowNames = Object.keys(selectedNode).filter((key) => typeof selectedNode[key] !== 'object'
     && !key.includes('label')).sort()
@@ -24,7 +24,7 @@ const NodesSelectionDetails = ({
   return (
     <div className="nodes-selection-details m-t-10">
       <h3 className="">
-        {`${t('node')}: ${availableNodesNormalised[nodeId].label}`}
+        {`${t('node')}: ${availableNodes.get(nodeId).label}`}
       </h3>
       <div className="nodes-selection-details-title">
         {t('properties')}
@@ -106,15 +106,15 @@ const NodesSelectionDetails = ({
 
 NodesSelectionDetails.propTypes = {
   nodeId: PropTypes.string.isRequired,
-  availableNodesNormalised: PropTypes.shape().isRequired,
+  availableNodes: PropTypes.shape().isRequired,
   nodesConnections: PropTypes.shape().isRequired,
 }
 
 const mapToProps = ({
-  availableNodesNormalised,
+  availableNodes,
   nodesConnections
 }) => ({
-  availableNodesNormalised,
+  availableNodes,
   nodesConnections
 })