+++ src/utils/serialiseNodesEdges/addNode.js
@@ -4,8 +4,7 @@ import { SELECTED_NODE_COLOR } from '../../constants/graph'
  * Add node to arrays/objects
  * @param  {Object}   params
  * @param  {Array}    params.addedNodes               Array of nodes IDs being added
- * @param  {Array}    params.availableNodesList       Array of available nodes
- * @param  {Object}   params.availableNodesNormalised Normalised list of available nodes
+ * @param  {Object}   params.availableNodes           Available nodes dataset
  * @param  {Array}    params.highlightedNodes         Array of nodes IDs to highlight
  * @param  {Boolean}  params.isNodeOverlay            Flag to make non-highlighted nodes transparent
  * @param  {String}   params.nodeId                   Node ID
@@ -15,8 +14,7 @@ import { SELECTED_NODE_COLOR } from '../../constants/graph'
  */
 const addNode = ({
   addedNodes,
-  availableNodesList,
-  availableNodesNormalised,
+  availableNodes,
   isNodeOverlay,
   highlightedNodes,
   nodeId,
@@ -24,6 +22,7 @@ const addNode = ({
   shortestPathNodes,
 }) => {
   if (!addedNodes.includes(nodeId)
+  && !availableNodes.get(nodeId)
   && nodeIdObject.label
   && nodeIdObject.label !== ''
   ) {
@@ -37,8 +36,7 @@ const addNode = ({
       extendedNodeObject.opacity = 0.1
     }
 
-    availableNodesNormalised[nodeId] = extendedNodeObject  // eslint-disable-line
-    availableNodesList.push(extendedNodeObject)
+    availableNodes.add(extendedNodeObject)
     addedNodes.push(nodeId)
   }
 }